data/ 植生調査票(2020-2025) VT Vegtaion cross table for the period V (2020-2025)

data/ 植生調査票(2020-2025) VT Vegtaion cross table for the period V (2020-2025)

data(VT)

Format

An object of class `list` of `tbl_df`.

Source

automatically generated skeleton

See also

Examples

names(VT)
#>  [1] "Bijodaira_dk01"     "Bijodaira_c02"      "Bijodaira_c03"     
#>  [4] "Bijodaira_c04"      "Bijodaira_c05"      "Bunazaka_dk01"     
#>  [7] "Bunazaka_c02"       "Bunazaka_c03"       "Bunazaka_c04"      
#> [10] "Bunazaka_c05"       "Bunadaira_dk01"     "Bunadaira_c02"     
#> [13] "Bunadaira_c03"      "Bunadaira_c04"      "Bunadaira_c05"     
#> [16] "Kaminokodaira_dk01" "Kaminokodaira_c02"  "Kaminokodaira_c03" 
#> [19] "Kaminokodaira_c04"  "Kaminokodaira_c05"  "Midagahara_dk01"   
#> [22] "Midagahara_c02"     "Midagahara_c03"     "Midagahara_c04"    
#> [25] "Midagahara_c05"     "Mimatsu_dk01"       "Mimatsu_c02"       
#> [28] "Mimatsu_c03"        "Mimatsu_c04"        "Mimatsu_c05"       
#> [31] "Matsuotoge_dk01"    "Matsuotoge_c02"     "Matsuotoge_c03"    
#> [34] "Matsuotoge_c04"     "Matsuotoge_c05"     "Kagamiishi_dk01"   
#> [37] "Kagamiishi_c02"     "Kagamiishi_c03"     "Kagamiishi_c04"    
#> [40] "Kagamiishi_c05"     "Joudosan_dk01"      "Joudosan_c02"      
#> [43] "Joudosan_c03"       "Joudosan_c04"       "Joudosan_c05"      
#> [46] "Arimine_dk01"       "Arimine_c02"        "Arimine_c03"       
#> [49] "Arimine_c04"        "Arimine_c05"       
 VT[["Arimine_c04"]]
#> # A tibble: 83 × 30
#>    sp      `1`   `2`   `3`   `4`   `5`   `6`   `7`   `8`   `9`  `10`  `11`  `12`
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 アオダモ…   1     3.5   0.5   3     0     0     1     1     0     0     1     3  
#>  2 イタヤカ…   0.3   0     0     0     0.5   0.5   0.5   0     0     0     0.5   0.5
#>  3 ウリハダ…   0.5  13     1.5   2    10     1     6     4     1     2     1     2  
#>  4 ウワミズ…   0.8   3     3     3     4     2     4     0.5   2     5     0     1  
#>  5 オオカメ…  14    37    30    20    15    25    23    35    21     4    20    20  
#>  6 オオバク…  10    25    20     4     6    18    15    19     7     7    13     8  
#>  7 コシアブ…   4     0     5.5   1.5   0.7   0     1     4     0     2     3     4  
#>  8 コマユミ…   0.1   0.5   0     0.5   0.5   0.5   0.5   0     0.5   0     0.2   0.2
#>  9 サワフタ…   2     0     0     0     0     0     0     1.5   7     0.5   0     0  
#> 10 タムシバ…   3     0     0     0     0     0     0     0.5   0.5   0     0     0  
#> # ℹ 73 more rows
#> # ℹ 17 more variables: `13` <dbl>, `14` <dbl>, `15` <dbl>, `16` <dbl>,
#> #   `17` <dbl>, `18` <dbl>, `19` <dbl>, `20` <dbl>, `21` <dbl>, `22` <dbl>,
#> #   `23` <dbl>, `24` <dbl>, `25` <dbl>, Mean <dbl>, Frequency <dbl>,
#> #   FreqCode <chr>, DominatIndex <dbl>
 VTdk[["Arimine_c04"]]
#> # A tibble: 83 × 30
#>    sp      `1`   `2`   `3`   `4`   `5`   `6`   `7`   `8`   `9`  `10`  `11`  `12`
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 アオダモ…   1     3     0.5   3     0     0     1     1     0     0     1     3  
#>  2 イタヤカ…   0.5   0     0     0     0.5   0.5   0.5   0     0     0     0.5   0.5
#>  3 ウリハダ…   0.5   5     1     2     5     1     4     3     1     2     1     2  
#>  4 ウワミズ…   0.5   3     3     3     3     2     3     0.5   2     4     0     1  
#>  5 オオカメ…   5     7     6     6     5     6     6     7     6     3     6     6  
#>  6 オオバク…   5     6     6     3     4     5     5     5     4     4     5     4  
#>  7 コシアブ…   3     0     4     1     0.5   0     1     3     0     2     3     3  
#>  8 コマユミ…   0.5   0.5   0     0.5   0.5   0.5   0.5   0     0.5   0     0.5   0.5
#>  9 サワフタ…   2     0     0     0     0     0     0     1     4     0.5   0     0  
#> 10 タムシバ…   3     0     0     0     0     0     0     0.5   0.5   0     0     0  
#> # ℹ 73 more rows
#> # ℹ 17 more variables: `13` <dbl>, `14` <dbl>, `15` <dbl>, `16` <dbl>,
#> #   `17` <dbl>, `18` <dbl>, `19` <dbl>, `20` <dbl>, `21` <dbl>, `22` <dbl>,
#> #   `23` <dbl>, `24` <dbl>, `25` <dbl>, Mean <dbl>, Frequency <dbl>,
#> #   FreqCode <chr>, DominatIndex <dbl>